refactor: 使用 jinja 来渲染加载测试脚本并调整运行方法
