+++ pyproject.toml
@@ -47,7 +47,7 @@ show-bump = "bump-my-version show-bump"
 snapshot-create = "pytest --inline-snapshot=create"
 snapshot-fix = "pytest --inline-snapshot=fix"
 store-test = "python -m src.providers.store_test"
-plugin-test = "python src/providers/docker_test/plugin_test.py"
+docker-test = "python -m src.providers.docker_test"
 
 [tool.pyright]
 pythonVersion = "3.12"
